fix(dashboard): Postgres int4/int8/numeric-Decode in Analytics-Handlern
Systemischer Port-Fehler: INTEGER/BIGINT/numeric-Aggregate wurden als
f64/i64 dekodiert, sqlx wirft ColumnDecode, der per .ok()/unwrap_or(0)
still verschluckt wurde -> Felder dauerhaft 0. Betroffen: category-
comparison/leaderboard (avg/peak), audience-demographics (viewerMinutes),
raid-retention/analytics (viewers_sent, chatter-Metriken, incoming-Boost),
rankings growth (SUM follower_delta), follower-funnel (total_duration),
monthly-stats (peak_viewers, airtime), session-detail-Fallback (chatters).

Fix: explizite SQL-Casts ::float8/::bigint am SELECT bzw. i32-Decode +
Widening, passend zum tatsaechlichen Spaltentyp. Verhalten jetzt = Python
(untypisiertes int()/float()).

+++ rust/crates/tb-dashboard-api/src/handlers/audience_demographics.rs
@@ -302,7 +302,7 @@ pub async fn audience_demographics_handler(
 
     // ── Q5: Viewer-Sample (twitch_session_viewers) ────────────────────────────
     let vsamp = sqlx::query(
-        "SELECT COUNT(*) AS cnt, COALESCE(SUM(GREATEST(sv.viewer_count,0)),0) AS vm
+        "SELECT COUNT(*) AS cnt, COALESCE(SUM(GREATEST(sv.viewer_count,0)),0)::float8 AS vm
          FROM twitch_session_viewers sv
          JOIN twitch_stream_sessions s ON s.id = sv.session_id
          WHERE s.started_at >= $1 AND LOWER(s.streamer_login) = $2 AND s.ended_at IS NOT NULL"
